requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.2
reportlab==4.2.2
python-dotenv==1.0.1
pydantic-settings==2.5.2
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
ruff==0.6.8
black==24.8.0
gunicorn==23.0.0
